+++ bot/commands/fun/cat.js
@@ -1,9 +1,8 @@
 'use strict';
-
 var request = require('request');
 
 module.exports = function(bot) {
-  request('http://random.cat/meow', function (error, response, body) {
+  request('http://aws.random.cat/meow', function (error, response, body) {
     if (!error && response.statusCode === 200) {
       var json = JSON.parse(body);
       bot.sendChat(json.file);