+++ CHANGELOG.md
@@ -7,6 +7,18 @@ The Commands of the bot are its API so here's how it will adhere to Semantic Ver
 **MINOR** - Adding new commands or functionality that won't affect or break existing ones    
 **PATCH** - Bug fixes, code improvements (backwards-compatible bug fixes)
 
+## [1.23.1] - 2018-05-14
+### Fixed
+- updated API for random.cat that `!cat` uses.  Closes [issue 45](https://github.com/FranciscoG/DerpyBot/issues/45)
+- bot will skip broken soundcloud links
+- removed the "recently warned" check, it will continually warn now
+
+### Added
+- added soundcloud.js utility tests
+
+### Changed
+- cleaned up some of the code
+
 ## [1.23.0] - 2018-05-01
 ### Fixed
 - fix issue with triggers not updating properly when changed